Extract useVideoNavigation hook from Home
