+++ frontend/src/hooks/frontend/useVideoNavigation.js
@@ -1,0 +1,26 @@
+import { useEffect } from 'react'
+import { useNavigate, useSearchParams } from 'react-router-dom'
+
+export function useVideoNavigation(id, videos, playlists) {
+    const navigate = useNavigate()
+    const [searchParams] = useSearchParams()
+
+    useEffect(() => {
+        if (!id || (videos.length > 0 && !videos.find(v => v.id === id))) {
+            const firstVideo = videos[0]
+            if (firstVideo) {
+                const playlist = playlists.find(p => p.some(v => v.id === firstVideo.id))
+                const firstId = playlist ? playlist[0].id : firstVideo.id
+                navigate(`/watch/${firstId}`, { replace: true })
+            }
+            return
+        }
+        if (!searchParams.get('t')) {
+            const savedTime = Math.floor(parseFloat(localStorage.getItem(`time_${id}`)))
+            if (savedTime && savedTime > 10) {
+                navigate(`/watch/${id}?t=${savedTime}`, { replace: true })
+                return
+            }
+        }
+    }, [id, videos, playlists])
+}
